test: refactor benchmark metric logic (#157)
- benchmark_metrics를 통한 공용 측정 모듈 생성
- 테스트 후 요약표, 케이스별 hit/miss, 목표 달성 여부, 검색 적중 여부 진단
- 측정 로직을 공용 모듈에서 import 하도록 변경
